+++ Kernel/include/sounds.h
@@ -6,6 +6,6 @@
 void beep();
 void play_sound(uint32_t nFrequence);
 void nosound();
-void playPiano(uint8_t frec);
+void playPiano(uint32_t frec);
 
 #endif
